Cleaned up code, fixed bugs
Clicking and dragging on the ball now works as intended. One thing I noticed is that you can click and drag when the ball is moving, wait, and then let go after the ball stops. We might want to fix this but I'm not sure how.

Working on implementing a visual indicator of the dragging as well as a maximum velocity.

Still can't get the ball to stop quickly, fine-tuning will come later.