Free Play now returns to Main Menu
Free play is enabled by the main menu free play button and disabled by the "back" button in the free play menu as well as the "back to main menu" button in the pause menu. I also adjusted the default value of the enter speed for the hole and removed some unneeded debug text.